@@ -11,45 +11,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.13",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.2.13",
-    "@angular/compiler": "~7.2.13",
-    "@angular/core": "~7.2.13",
-    "@angular/forms": "~7.2.13",
-    "@angular/material": "^7.3.6",
-    "@angular/platform-browser": "~7.2.13",
-    "@angular/platform-browser-dynamic": "~7.2.13",
-    "@angular/router": "~7.2.13",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.1",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^4.2.2",
+    "bootstrap": "^4.4.1",
+    "core-js": "^2.6.11",
     "hammerjs": "^2.0.8",
-    "ol": "^5.3.1",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "ol": "^5.3.3",
+    "rxjs": "~6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.13",
-    "@angular/language-service": "~7.2.13",
+    "@angular-devkit/build-angular": "~0.803.21",
+    "@angular/cli": "^8.3.21",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^13.1.2",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "prettier": "1.17.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   }
 }
